==> workflow/01-data.Rmd <==
---
title: "01-data"
date: "Compiled at `r format(Sys.time(), '%Y-%m-%d %H:%M:%S', tz = 'UTC')` UTC"
output: github_document
params:
  name: "01-data" # change if you rename file
---

```{r here, message=FALSE,echo=F}
here::i_am(paste0(params$name, ".Rmd"), uuid = "b3cd4969-aab1-4ff2-a553-9df4269240e4")
```

The purpose of this document is to introduce three flow cytometry data that we'll use for clustering comparison. They are:

-   The Z-project data

-   The B-sub punch biopsy data

```{r packages, message=F}
library("conflicted")
library(purrr)
library(dplyr)
library(flowCore)
library(flowWorkspace)
library(ggcyto)
library(tidyverse)
library(gridExtra)
library(knitr)
library(ggplot2)
library(RColorBrewer)
```

```{r directories,echo=F}
# create or *empty* the target directory, used to write this file's data: 
projthis::proj_create_dir_target(params$name, clean = TRUE)

# function to get path to target directory: path_target("sample.csv")
path_target <- projthis::proj_path_target(params$name)

# function to get path to previous data: path_source("00-import", "sample.csv")
path_source <- projthis::proj_path_source(params$name)
```

## Z-project

Z-project is focusing on a colony of Bacillus subtilis, lab exams the sub-population structure at five different locations: inner zone, middle zone, outer zone, surrounding and whole colony. 

```{r import-z}
# import data
fcs.direction <- "~/Desktop/Z-Project_Bacillus_pretest/fcs/"
fcs.file.name <- list.files(fcs.direction,pattern = "\\.fcs$",full.names = T)
z_fs <- list()

for (i in fcs.file.name){
  data <- read.FCS(i,alter.names = T,transformation = F)
  new_name <- substr(i,start = 55,stop=nchar(i))
  z_fs[[new_name]] <- data
}
```

```{r explanation-z,echo=T}
markernames(z_fs[[2]])
```

DAPI, FDA are common fluorescent dyes which are excited with 488nm and measured with 460/50nm, 530/40nm respectively. 

Based on above table, the relationship between fluorescent dyes and variables are: DAPI-PMT.9, FDA-PMT.3.

We regard cells with measurement results below 200 on interested channels, namely FSC, SSC, DAPI/FDA_PI, as debris and remove them from the data to ensure the accuracy of downstream analysis

```{r debris-z}
location <- c("Inner_zone","Middle_zone","Outer_zone","Surrounding","Whole_colony")

# remove the debris
new_DAPI_fs <- list()
new_FDA_PI_fs <- list()

for (i in 1:5) {
  data_name <- paste0(location[i],"_DAPI.fcs")
  data <- z_fs[[data_name]]
  sub <- data@exprs %>% as.data.frame() %>% dplyr::filter(PMT.1>200) %>% dplyr::filter(PMT.2>200)  %>% dplyr::filter(PMT.9>200)
  sub <- as.matrix(sub)
  data@exprs <- sub
  new_DAPI_fs[[data_name]]<-data
}

for (i in 1:5) {
  data_name <- paste0(location[i],"_FDA_PI.fcs")
  data <- z_fs[[data_name]]
  sub <- data@exprs %>% as.data.frame() %>% dplyr::filter(PMT.1>200) %>% dplyr::filter(PMT.2>200) %>% dplyr::filter(PMT.3>200)
  sub <- as.matrix(sub)
  data@exprs <- sub
  new_FDA_PI_fs[[data_name]]<-data
}
```

```{r save-z,eval=F,message=F}
write_rds(new_DAPI_fs,"~/Desktop/new_DAPI.rds")
write_rds(new_FDA_PI_fs,"~/Desktop/new_FDA_PI.rds")
```

Now, let's have a look of cytograms from Z-project.
```{r plot-z}
for (i in 1:5){
  data_name <- paste0(location[i],"_DAPI.fcs")
  data <- new_DAPI_fs[[data_name]]
  p <- ggcyto(data, aes(x = "PMT.1", y = "PMT.9")) + geom_hex(bins = 128)
  print(p)
}

for (i in 1:5){
  data_name <- paste0(location[i],"_FDA_PI.fcs")
  data <- new_FDA_PI_fs[[data_name]]
  p <- ggcyto(data, aes(x = "PMT.1", y = "PMT.3")) + geom_hex(bins = 128)
  print(p)
}
```

This is another visulization method offered by flowEMMI v2 packages.
```{r,echo=T,eval=F}
# replace the fcsData and ch1/2
p <- plotDensityAndEllipses(fcsData = data, ch1="PMT.1", ch2="PMT.3",
                          logScale = F, title = data_name,
                          axis_size=10, axisLabeling_size=10,
                          xlab = "Forward Scatter", ylab = "Fluorescence", 
                          font = "Arial")
```

## B-sub punch biopsy

```{r import-punch}
# import data
fcs.direction <- "~/Desktop/B-sub/"
fcs.file.name <- list.files(fcs.direction,pattern = "\\.fcs$",full.names = T)
B_punch_fs <- list()

for (i in fcs.file.name){
  data <- read.FCS(i,alter.names = T,transformation = F)
  new_name <- substr(i,start = 41,stop=nchar(i))
  B_punch_fs[[new_name]] <- data
}
```

The punch biopsy method uses 4 different-sized needles to sample mono-species biofilms from the inside out. The stain is Syto9, which can guarantee the cell viability. And this kind of stain is excited at 530/40 nm level, correlated to variable PMT.3.

```{r debris-punch}
new_B_punch_fs <- list()

for (i in 1:22) {
  data_name <- names(B_punch_fs)[i]
  data <- B_punch_fs[[data_name]]
  sub <- data@exprs %>% as.data.frame() %>% dplyr::filter(PMT.1>200) %>% dplyr::filter(PMT.2>200)  %>% dplyr::filter(PMT.3>200)
  sub <- as.matrix(sub)
  data@exprs <- sub
  new_B_punch_fs[[data_name]]<-data
}

```

```{r plot-punch}

for (i in 1:22){
  data_name <- names(new_B_punch_fs)[i]
  data <- new_B_punch_fs[[data_name]]
  p <- ggcyto(data, aes(x = "PMT.1", y = "PMT.3")) + geom_hex(bins = 128)
  print(p)
}


```


## Files written

These files have been written to the target directory, `r paste0("data/", params$name)`:

```{r list-files-target}
projthis::proj_dir_info(path_target())
```
