==> workflow/04a-GMM for Z-project.Rmd <==
---
title: "04a-GMM for Z-project"
date: "Compiled at `r format(Sys.time(), '%Y-%m-%d %H:%M:%S', tz = 'UTC')` UTC"
output: github_document
params:
  name: "04a-GMM for Z-project" # change if you rename file
---

```{r here, message=FALSE}
here::i_am(paste0(params$name, ".Rmd"), uuid = "9d6a84ea-ebe8-4073-bfe1-9e2529a9d667")
```

The purpose of this document is ...

```{r packages}
library("conflicted")
library(purrr)
library(dplyr)
library(flowCore)
library(ggplot2)
library(mclust)
library(flowEMMi)
library(factoextra)
library(flowClust)
library(flowViz)
library(flowMerge)
library(MASS)
library(ggpubr)
library(plot3D)
library(VGAMextra)
library(mvtnorm)
```

```{r directories}
# create or *empty* the target directory, used to write this file's data: 
projthis::proj_create_dir_target(params$name, clean = TRUE)

# function to get path to target directory: path_target("sample.csv")
path_target <- projthis::proj_path_target(params$name)

# function to get path to previous data: path_source("00-import", "sample.csv")
path_source <- projthis::proj_path_source(params$name)
```

## Import data
```{r import}
DAPI <- readRDS("~/Desktop/MSc_new_data/new_DAPI.rds")
FDA_PI <- readRDS("~/Desktop/MSc_new_data/new_FDA_PI.rds")
gating_DAPI <- readRDS("~/Desktop/MSc_new_data/gating_DAPI.rds")
gating_FDA_PI <- readRDS("~/Desktop/MSc_new_data/gating_FDA_PI.rds")
```

Generate function body.

```{r function}
flowGMM <- function(data,n_cluster){
  model1 <- Mclust(data,G=n_cluster,modelNames = "VVV")
  summary(model1)

  clusters1 <- predict(model1)
  result1 <- cbind(data,Cluster=clusters1$classification)
  result1 <- as.data.frame(result1)
  result1$Cluster <- as.factor(result1$Cluster)
  
  result <- list(model=model1,parameter=model1$parameters,class_result=result1)
  return(result)
}
```

# GMM 

### Gating

In the first part, we don't specify the number of clusters, the best results will be determined by cross validation.

```{r GMM-gating,eval=FALSE}
DAPI_GMM <- list()
DAPI_GMM_flowemmi <- list()

for (i in 1:5){
  data <- DAPI[[i]]@exprs[,c(11,27)]
  data_name <- names(DAPI)[i]
  DAPI_GMM[[data_name]]<-flowGMM(data,NULL)
  
  n_cluster <- length(gating_DAPI[[i]]@sigma)
  DAPI_GMM_flowemmi[[data_name]]<-flowGMM(data,n_cluster)
}
```

```{r,echo=FALSE}
DAPI_GMM_flex <- readRDS("~/Desktop/MSc_new_data/DAPI_GMM_flex.rds")
DAPI_GMM_fix <- readRDS("~/Desktop/MSc_new_data/DAPI_GMM_fix.rds")

```

### gating plot
```{r GMM-plots}
DAPI_GMM_3D <- list()
GMM_plot <- list()

for (i in 1:5){
  data_name <- names(DAPI)[i]
  GMM_plot[[i]] <- fviz_mclust(DAPI_GMM_flex[[i]]$model, "classification", geom = "point",main=data_name)
  GMM_plot[[i+5]] <- fviz_mclust(DAPI_GMM_fix[[i]]$model, "classification", 
                                 geom = "point",main="K = flowEMMI")
}

# 2D plots
ggarrange(plotlist = GMM_plot,ncol=5,nrow=2,common.legend = T,legend = "bottom")

# 3D plots
for (i in 1:5){
  data1 <- DAPI[[i]]@exprs[,c(11,13,27)]
  data_name <- names(DAPI)[i]
  DAPI_GMM_3D[[data_name]]<-flowGMM(data1,NULL)
  plot(DAPI_GMM_3D[[i]]$model,what="classification",main=data_name)
}

```

### Meta-clustering 

```{r GMM-merge}
GMM.merge <- function(gating_list){
  
 all_gate <- gating_DAPI[[1]]
 all_gate@mu <- matrix(NA,nrow=2,ncol=1)
 all_gate@sigma <- list()
 all_gate@clusterProbs <- numeric()

 for (i in 1:5){
    mu <- gating_list[[i]]$parameter$mean
    dim <- dim(gating_list[[i]]$parameter$variance$sigma)[3]
  
    sigma <- list()
    for (j in 1:dim){
      sigma[[j]] <- gating_list[[i]]$parameter$variance$sigma[,,j]
    }
    
    prob <- gating_list[[i]]$parameter$pro
    
    all_gate@mu <- cbind(all_gate@mu,mu)
    all_gate@sigma <- c(all_gate@sigma,sigma)
    all_gate@clusterProbs <- c(all_gate@clusterProbs,prob)
 }
 
 all_gate@mu <- all_gate@mu[,-1]
 
 meta_result <- removeOverlaps(em=all_gate, alpha=0.9
                                  , mergeWhenCenter = FALSE
                                  , mergeWhenTwoCenters = FALSE
                                  , thresholdForDeletion = 0.5
                                  , threshold = 0.9
                                  , shrinkingFunction=shrinkEllipses
                                  , considerWeights=TRUE
                                  , plot = FALSE
                                  , minMinor = 500)
 return(meta_result)
}

```

### meta-gating plots

```{r GMM-meta-plot}
location <- c("Inner_zone","Middle_zone","Outer_zone","Surrounding","Whole_colony")
GMM_flex_meta <- GMM.merge(DAPI_GMM_flex)
GMM_fix_meta <- GMM.merge(DAPI_GMM_fix)

meta_plots <- list()

for (i in 1:5){
  data <- DAPI[[i]]
  var <- location[i]
  meta_plots[[i]] <- plotDensityAndEllipses(fcsData = data, ch1="PMT.1", ch2="PMT.9", alpha=0.9,
                            logScale = F, results = GMM_flex_meta,
                            title = paste0("GMM Meta-clust on ",var," with flex. K"), 
                            plotRelevance = T,gridsize = 1000,
                            ellipseDotSize = 0.5, axis_size=10, axisLabeling_size=10,
                            xlab = "Forward Scatter", ylab = "DAPI", font = "Arial")
}

```

## ROC Function

### for points

```{r roc-function}
GMMroc <- function(data,gating_data,dimension,CI){
  mu <- gating_data$parameter$mean
  dim <- dim(gating_data$parameter$variance$sigma)[3]
  
  sigma <- list()
  for (i in 1:dim){
    sigma[[i]] <- gating_data$parameter$variance$sigma[,,i]
  }
  
  names <- colnames(data)
  
  ncells <- nrow(data)
  n_clusters <- length(sigma)
  
  #generate mahalanobis matrix
  maha_data <- matrix(NA,nrow=ncells,ncol=n_clusters)
  
  for (i in 1:ncells){
    for(j in 1:n_clusters){
      maha_data[i,j] <- mahalanobis(data[i,],mu[,j],sigma[[j]])
    }
  }
  
  maha <- maha_data[,2:n_clusters] %>% as.data.frame()
  
  data1 <- maha
  
  #randomly sample points
  set.seed(123)
  num_points <- 100000
  
  random_points <- runif(num_points*dimension,min=200,max=65535) %>% round(2)
  random_data <- matrix(random_points,ncol=dimension,nrow=num_points)
  
  #maha matrics for random points
  random_maha <- matrix(NA,nrow=num_points,ncol=n_clusters)
  for (i in 1:num_points){
    for(j in 1:n_clusters){
      random_maha[i,j] <- mahalanobis(random_data[i,],mu[,j],sigma[[j]])
    }
  }
  
  random_maha <- as.data.frame(random_maha)
  roc.points <- data.frame()
  
  #performance at diff. alpha
  for (alpha in CI){
    threshold <- qinv.chisq(alpha,df=dimension)
    
    # cell
    for (cell in 1:ncells){
      rv <- data1[cell,1:ncol(data1)-1]
      if(all(rv>threshold)) {data1$Cluster[cell] <- NA}
      else {data1$Cluster[cell] <- which.min(rv)}
    }
  
    result1 <- base::table(data1$Cluster) %>% as.data.frame()
    sum1 <- sum(result1$Freq)
    
    # volume
    for (cell in 1:num_points){
      rv <- random_maha[cell,1:ncol(random_maha)-1]
      if(all(rv>threshold)) {random_maha$Cluster[cell] <- NA}
      else {random_maha$Cluster[cell] <- which.min(rv)}
    }
  
    result2 <- base::table(random_maha$Cluster) %>% as.data.frame()
    sum2 <- sum(result2$Freq)/num_points*(65535-200)^dimension
    
    
    results <- cbind(cell=sum1,volume=sum2) %>% as.data.frame()
    roc.points <- rbind(roc.points,results)
  }
  result <- list(maha.matrix=data1,roc.table=roc.points)
  return(result)
}
```

### for plots

```{r roc-plot}
# plot
plot.roc <- function(roc_result){
  
  plot.points <- data.frame()
  plot.scale <- data.frame()
  plot.scale.points <- data.frame()

  for (i in 1:5){
    data <- roc_result[[i]]
    data[1,] <- c(0,0)
  
  # No Scaling
  data$location <- substr(names(DAPI)[i],1,nchar(names(DAPI)[i])-9) %>% as.factor()
  plot.points <- rbind(plot.points,data)
  
  # Scaled by regional maximum
  max <- data[100,]
  for (j in 1:100){
    data[j,1] <- data[j,1]/max[1,1]
    data[j,2] <- data[j,2]/max[1,2]
  }
  plot.scale <- rbind(plot.scale,data1)
  
  
  # Scaled by the max number among all regions
  plot.scale.points <- plot.points
  max1 <- max(plot.scale.points[,1])
  max2 <- max(plot.scale.points[,2])
  
  for (j in 1:nrow(plot.points)){
    plot.scale.points[j,1]<- plot.scale.points[j,1]/max1
    plot.scale.points[j,2]<- plot.scale.points[j,2]/max2
    }
  }
  
  points <- list(raw.point=plot.points,regional.scaled=plot.scale,all.scaled=plot.scale.points)
  return(points)
  
}
```

### Compute ROC points

Using different gating results

```{r roc-calc points,eval=FALSE}
GMM.roc <- list()

for (i in 1:5){
  data <- DAPI[[i]]@exprs[,c(11,27)]
  data_name <- names(DAPI)[i]
  gating_data <- DAPI_GMM_flowemmi[[i]]
  alpha <- seq(0.01,0.99,0.01)
  GMM.roc[[data_name]] <- GMMroc(data,gating_data,alpha)
}
```

# flowClust

### gating

```{r flowClust-gating}
DAPI_fC_flex <- list()
DAPI_fC_fixed <- list()

# gating
for (i in 1:5){
  res <- flowClust(DAPI[[i]], varNames = c("PMT.1","PMT.9"), K=1:10)
  data_name <- names(DAPI)[i]
  bic <- criterion(res, "BIC")
  max <- which.max(bic)
  plot(res[[max]], data = DAPI[[i]], main=paste0("Flex. result for ",data_name))
  DAPI_fC_flex[[data_name]] <- res[[max]]
  
  k <- length(gating_DAPI[[i]]@sigma)
  DAPI_fC_fixed[[data_name]] <- res[[k]]
  plot(res[[k]], data = DAPI[[i]], main=paste0("Fixed result for ",data_name))
}

```

### flowMerge
```{r flowClust-merge}
DAPI_fC_flex_meta <- list()
DAPI_fC_fixed_meta <- list()

for (i in 1:5){
  data <- DAPI[[i]]
  data_name <- names(DAPI)[i]
  
  flowobj <- flowObj(DAPI_fC_flex[[i]],data)
  merge <- flowMerge::merge(flowobj,metric="mahalanobis")
  a <- fitPiecewiseLinreg(merge)
  
  flowobj2 <- flowObj(DAPI_fC_fixed[[i]],data)
  merge2 <- flowMerge::merge(flowobj2,metric="mahalanobis")
  b <-  fitPiecewiseLinreg(merge2)
  
  DAPI_fC_flex_meta[[data_name]] <- merge[[a]]
  DAPI_fC_fixed_meta[[data_name]] <- merge2[[b]]
}

#plots
for (i in 1:5){
  opt <- DAPI_fC_flex_meta[[i]]
  var <- location[i]
  plot(opt,level=0.9,main=paste0("Mering for flex K\nIn ",var))
}

```


## ROC Function

### for points

```{r flowClust-roc}
flowClust_roc <- function(data,gating_data,dimension,CI){
  
  # Box-Cox trans
  lambda <- gating_data@lambda
  data <- flowClust::box(data,lambda)
  
  # maha-matrix
  mu <- t(gating_data@mu)
  dim <- dim(gating_data@sigma)[1]
    
  sigma <- list()
  for (i in 1:dim){
    sigma[[i]] <- gating_data@sigma[i,,]
  }
  
  names <- colnames(data)
  
  ncells <- nrow(data)
  n_clusters <- length(sigma)
  
  #generate mahalanobis matrix
  maha_data <- matrix(NA,nrow=ncells,ncol=n_clusters)
  
  for (i in 1:ncells){
    for(j in 1:n_clusters){
      maha_data[i,j] <- mahalanobis(data[i,],mu[,j],sigma[[j]])
    }
  }
  
  data1 <- as.data.frame(maha_data)
  
  #randomly sample points
  set.seed(123)
  num_points <- 100000
  
  random_points <- runif(num_points*dimension,min=200,max=65535)
  random_boxcox <- flowClust::box(random_points,lambda) %>% round(2)
  random_data <- matrix(random_boxcox,ncol=dimension,nrow=num_points)
  
  #maha matrics for random points
  random_maha <- matrix(NA,nrow=num_points,ncol=n_clusters)
  for (i in 1:num_points){
    for(j in 1:n_clusters){
      random_maha[i,j] <- mahalanobis(random_data[i,],mu[,j],sigma[[j]])
    }
  }
  
  random_maha <- as.data.frame(random_maha)
  roc.points <- data.frame()
  
  #performance at diff. alpha
  for (alpha in CI){
    threshold <- dimension*qf(alpha,dimension,4)
    
    # cell
    for (cell in 1:ncells){
      rv <- data1[cell,1:ncol(data1)-1]
      if(all(rv>threshold)) {data1$Cluster[cell] <- NA}
      else {data1$Cluster[cell] <- which.min(rv)}
    }
  
    result1 <- base::table(data1$Cluster) %>% as.data.frame()
    sum1 <- sum(result1$Freq)
    
    # volume
    for (cell in 1:num_points){
      rv <- random_maha[cell,1:ncol(random_maha)-1]
      if(all(rv>threshold)) {random_maha$Cluster[cell] <- NA}
      else {random_maha$Cluster[cell] <- which.min(rv)}
    }
  
    result2 <- base::table(random_maha$Cluster) %>% as.data.frame()
    sum2 <- sum(result2$Freq)/num_points*(65535-200)^dimension
    
    
    results <- cbind(cell=sum1,volume=sum2) %>% as.data.frame()
    roc.points <- rbind(roc.points,results)
  }
    result <- list(maha.matrix=data1,roc.table=roc.points)
  return(result)
}
```

```{r flowClust1,eval=FALSE}
DAPI_fC_flex_roc <- list()
DAPI_fC_fix_roc <- list()

for (i in 1:5){
  name <- names(DAPI)[i]
  data0 <- DAPI[[name]]@exprs[,c(11,27)]
  gating_data <- DAPI_fC_fix[[name]]
  alpha <- seq(0.01,0.99,by=0.01)
  DAPI_fC_fix_roc[[name]] <- flowClust_roc(data0,gating_data,2,alpha)
}
```


## Files written

These files have been written to the target directory, ```r paste0("data/", params$name)```:

```{r list-files-target}
projthis::proj_dir_info(path_target())
```
